Trip Logs - Cash Transactions Revision
+Listing based on Status (request, approved, release)
+Added checkbox for trip selection on drivers summary
+Added filter on cash transactions
+Added trip logs on drivers on cash transactions
+Print form on cash transaction approved listings
+Print form remove navigation bar
+Added Company on Copy Trip

+++ JobsV1/Areas/Personel/Controllers/CarRentalCashReleaseController.cs
@@ -7,18 +7,74 @@ using System.Net;
 using System.Web;
 using System.Web.Mvc;
 using JobsV1.Areas.Personel.Models;
+using JobsV1.Models;
 
 namespace JobsV1.Areas.Personel.Controllers
 {
+
     public class CarRentalCashReleaseController : Controller
     {
         private CarRentalLogDBContainer db = new CarRentalLogDBContainer();
+        private DateClass dt = new DateClass();
 
         // GET: Personel/CarRentalCashRelease
-        public ActionResult Index()
+        public ActionResult Index(int? statusId)
         {
-            var crLogCashReleases = db.crLogCashReleases.Include(c => c.crLogDriver).Include(c => c.crLogClosing);
-            return View(crLogCashReleases.ToList());
+            var today = dt.GetCurrentDate();
+            var DateFilter = today.AddDays(-7);
+            if (statusId == null)
+                statusId = 1;
+
+            //get cash releases up to -7 days from today
+            var crLogCashReleases = db.crLogCashReleases.Include(c => c.crLogDriver).Include(c => c.crLogClosing)
+                .Where(c => DbFunctions.TruncateTime(c.DtRelease) > DateFilter);
+
+            List<crLogCashRelease> cashReleases = new List<crLogCashRelease>();
+
+            foreach (var log in crLogCashReleases.ToList())
+            {
+                var lateststatusId = getLatestStatusId(log.Id);
+
+                if (lateststatusId == statusId)
+                {
+                    //add request and accecpted logs
+                    if (log.DtRelease.Date <= today.Date && lateststatusId < 3)
+                        cashReleases.Add(log);
+                    //add returned logs with date today
+                    if (log.DtRelease.Date == today.Date && lateststatusId == 3)
+                        cashReleases.Add(log);
+                }
+            }
+
+            ViewBag.IsAdmin = User.IsInRole("Admin");
+            ViewBag.StatusId = statusId;
+
+            return View(cashReleases.ToList());
+        }
+
+
+        // GET: Personel/crLogFuels
+        public ActionResult PrevRecords()
+        {
+            var today = dt.GetCurrentDate();
+            var crLogCashReleases = db.crLogCashReleases.Include(c => c.crLogDriver).Include(c => c.crLogClosing)
+              .Where(c => DbFunctions.TruncateTime(c.DtRelease) < today).OrderByDescending(c => c.DtRelease);
+
+            List<crLogCashRelease> cashReleases = new List<crLogCashRelease>();
+
+            foreach (var log in crLogCashReleases.ToList())
+            {
+                var lateststatusId = getLatestStatusId(log.Id);
+
+                //add returned logs with date today
+                if (log.DtRelease.Date < today.Date && lateststatusId == 3)
+                    cashReleases.Add(log);
+                
+            }
+
+            ViewBag.IsAdmin = User.IsInRole("Admin");
+
+            return View(cashReleases.ToList());
         }
 
         // GET: Personel/CarRentalCashRelease/Details/5
@@ -121,10 +177,18 @@ namespace JobsV1.Areas.Personel.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            crLogCashRelease crLogCashRelease = db.crLogCashReleases.Find(id);
-            db.crLogCashReleases.Remove(crLogCashRelease);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (DeleteStatusRecords(id))
+            {
+                crLogCashRelease crLogCashRelease = db.crLogCashReleases.Find(id);
+                db.crLogCashReleases.Remove(crLogCashRelease);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+
         }
 
         public ActionResult DriverSummary(int id)
@@ -140,5 +204,241 @@ namespace JobsV1.Areas.Personel.Controllers
             }
             base.Dispose(disposing);
         }
+
+
+        public bool DeleteStatusRecords(int crLogCashReleaseId)
+        {
+            try
+            {
+                var logCashStatusList = db.crLogCashStatus.Where(c => c.crLogCashReleaseId == crLogCashReleaseId).ToList();
+
+                if (logCashStatusList != null)
+                {
+                    db.crLogCashStatus.RemoveRange(logCashStatusList);
+                    db.SaveChanges();
+                }
+
+                return true;
+            }
+            catch
+            {
+                return false;
+            }
+        }
+
+        // POST: Personel/CarRentalCashRelease/CreateDriverRelease
+        [HttpPost]
+        public bool CreateDriverRelease(DriverReleaseRequest releaseRequest)
+        {
+            try
+            {
+                var statusFlag = false;
+
+                crLogCashRelease crtrx = new crLogCashRelease();
+                crtrx.DtRelease = dt.GetCurrentDateTime();
+                crtrx.crLogDriverId = releaseRequest.DriverId;
+                crtrx.crLogClosingId = GenerateClosingId();
+                crtrx.Amount = releaseRequest.Amount;
+                crtrx.Remarks = releaseRequest.Remarks;
+
+                db.crLogCashReleases.Add(crtrx);
+                db.SaveChanges();
+
+                foreach (var tripId in releaseRequest.TripIds)
+                {
+                    statusFlag = UpdateTripLogClosingId(tripId, crtrx.crLogClosingId);
+                    if (statusFlag == false)
+                        return false;
+                }
+
+                //add status
+                AddLogStatus(crtrx.Id, 1);
+
+                return true;
+
+            }
+            catch
+            {
+                return false;
+            }
+        }
+
+        //Generate Closing Id
+        private int GenerateClosingId()
+        {
+            try
+            {
+                var today = dt.GetCurrentDateTime();
+
+                crLogClosing crLog = new crLogClosing();
+                crLog.dtClose = today;
+
+                db.crLogClosings.Add(crLog);
+                db.SaveChanges();
+
+                return crLog.Id;
+            }
+            catch 
+            {
+                return 0;
+            }
+        }
+
+        //Update TripLog ClosingId
+        private bool UpdateTripLogClosingId(int? id, int? closingId)
+        {
+            try
+            {
+                if (id == null || closingId == null)
+                    return false;
+
+                crLogTrip logTrip = db.crLogTrips.Find(id);
+                logTrip.crLogClosingId = closingId;
+
+                db.Entry(logTrip).State = EntityState.Modified;
+                db.SaveChanges();
+
+                return true;
+            }
+            catch
+            {
+                return false;
+            }
+        }
+
+        public int getLatestStatusId(int id)
+        {
+            var logStatusQuery = db.crLogCashStatus.Where(c => c.crLogCashReleaseId == id).OrderByDescending(c => c.Id).FirstOrDefault();
+
+            if (logStatusQuery != null)
+            {
+                return logStatusQuery.crCashReqStatusId;
+            }
+            else
+            {
+                //default 
+                return 1;
+            }
+        }
+
+
+        public bool AddLogStatus(int? id, int statusId)
+        {
+            try
+            {
+                if (id == null)
+                {
+                    return false;
+                }
+
+                //create new status log
+                crLogCashStatus logCashStatus = new crLogCashStatus();
+                logCashStatus.crCashReqStatusId = statusId;
+                logCashStatus.crLogCashReleaseId = (int)id;
+                logCashStatus.dtStatus = dt.GetCurrentDateTime();
+
+                db.crLogCashStatus.Add(logCashStatus);
+                db.SaveChanges();
+
+                return true;
+            }
+            catch
+            {
+                return false;
+            }
+        }
+
+        [HttpPost]
+        public bool ApproveRequest(int? id)
+        {
+            try
+            {
+                if (id == null)
+                    return false;
+                return AddLogStatus(id, 2);
+            }
+            catch
+            {
+                return false;
+            }
+        }
+
+        [HttpPost]
+        public bool ApproveRelease(int? id)
+        {
+            try
+            {
+                if (id == null)
+                    return false;
+                return AddLogStatus(id, 3);
+            }
+            catch
+            {
+                return false;
+            }
+        }
+
+
+        public ActionResult PrintApproveForm(int? id)
+        {
+
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+
+            var cashRelease = db.crLogCashReleases.Find(id);
+            var tripLogs = db.crLogTrips.Where(c => c.crLogClosingId == cashRelease.crLogClosingId).ToList();
+            if (cashRelease == null)
+            {
+                return HttpNotFound();
+            }
+            ViewBag.crLogTrips = tripLogs ?? new List<crLogTrip>();
+
+            return View(cashRelease);
+        }
+
+
+        public ActionResult DriverTripList(int? id)
+        {
+            string tripLogErr = "";
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+
+            var cashRelease = db.crLogCashReleases.Find(id);
+
+            var tripLogs = new List<crLogTrip>();
+            if (cashRelease.crLogClosingId != null)
+            {
+                tripLogs = db.crLogTrips.Where(c => c.crLogClosingId == cashRelease.crLogClosingId).ToList();
+            }
+            else
+            {
+                tripLogErr = "Closing Id is null";
+            }
+
+            if (cashRelease == null)
+            {
+                return HttpNotFound();
+            }
+            ViewBag.tripLogErr = tripLogErr;
+            ViewBag.DtRelease = cashRelease.DtRelease;
+            ViewBag.Driver = cashRelease.crLogDriver.Name;
+            ViewBag.Amount = cashRelease.Amount;
+            ViewBag.Remarks = cashRelease.Remarks;
+            return View(tripLogs);
+        }
+
     }
+}
+
+
+public class DriverReleaseRequest
+{
+    public int DriverId { get; set; }
+    public decimal Amount { get; set; }
+    public ICollection<int> TripIds { get; set; }
+    public string Remarks { get; set; }
 }
